git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/qfsm/main@3112 0bcb941b-74d1-0310-bb26-e6636707f451

+++ prep/fsmPrepMain.m
@@ -347,7 +347,7 @@ for counter1=1:n
                rBgMaskInd = find(bgMaskFileIndex==currentIndex);
 
                if ~isempty(rBgMaskInd)
-                  bwMask = imread(bgMaskFileList{rBgMaskInd});
+                  bwMask = double(imread(bgMaskFileList{rBgMaskInd}));
                   successCE = 1;
                end
             end
